requirements.txt

numpy == 1.24.3
pandas == 2.0.1
psycopg2-binary-2.9.6
loguru==0.7.0

alembic==1.7.6
fastapi==0.73.0
psycopg2-binary==2.9.3
uvicorn==0.16.0
requests

catboost==1.1.1

tqdm==4.64.1